uriloader/exthandler/tests/mochitest/handlerApp.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <title>Pseudo Web Handler App</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="onLoad()">
Pseudo Web Handler App

<script class="testbody" type="text/javascript">
<![CDATA[
function onLoad() {

  // if we have a window.opener, this must be the windowContext
  // instance of this test.  check that we got the URI right and clean up.
  if (window.opener) {
    window.opener.is(location.search, 
                     "?uri=" + encodeURIComponent(window.opener.testURI), 
                     "uri passed to web-handler app");
    window.opener.SimpleTest.finish();
  } 

  window.close();
}
]]>
</script>

</body>
</html>
